requirements.txt
openai>=1.52.0
langfuse>=2.0.0
pydantic>=2.7.0
pytest>=8.0.0
python-dotenv>=1.0.0
